Smaller changes related to memory management.
Params and DictionaryFactory are now lazy Meyers singletons that return references. Calls and such adjusted accordingly.
This prevents a previous problem where the destructor was not called for the singletons, since it used a pointer
_instance.
In several instances, pointers would not be properly delete'd (AlignMake hypothesis, WordToken sequences). Fixed that
too.